+++ src/us/kbase/assemblyhomology/minhash/MinHashImplementationFactory.java
@@ -14,10 +14,11 @@ public interface MinHashImplementationFactory {
 	
 	/** Get the implementation from this factory.
 	 * @param tempFileDirectory a directory the implementation can use to create temporary files.
+	 * @param minhashTimeoutSec the timeout to use for the minhash process in seconds.
 	 * @return the implementation.
 	 * @throws MinHashInitException if the implementation could not be created.
 	 */
-	MinHashImplementation getImplementation(final Path tempFileDirectory)
+	MinHashImplementation getImplementation(Path tempFileDirectory, int minhashTimeoutSec)
 			throws MinHashInitException;
 	
 	/** Get the name of the implementation.