+++ src/us/kbase/assemblyhomology/service/exceptions/ErrorMessage.java
@@ -92,49 +92,49 @@ public class ErrorMessage {
 	}
 
 	/** Get the HTTP code for the error message.
-	 * @return
+	 * @return the HTTP code.
 	 */
 	public int getHttpcode() {
 		return httpcode;
 	}
 
 	/** Get the HTTP status string for the error message.
-	 * @return
+	 * @return the HTTP status.
 	 */
 	public String getHttpstatus() {
 		return httpstatus;
 	}
 
 	/** Get the application code for the error message.
-	 * @return
+	 * @return the application code.
 	 */
 	public Integer getAppcode() {
 		return appcode;
 	}
 
 	/** Get the application error string for the error message
-	 * @return
+	 * @return the application error.
 	 */
 	public String getApperror() {
 		return apperror;
 	}
 
 	/** Get the error message.
-	 * @return
+	 * @return the error message.
 	 */
 	public String getMessage() {
 		return message;
 	}
 
 	/** Get the call ID under which the error occurred.
-	 * @return
+	 * @return the call ID.
 	 */
 	public String getCallid() {
 		return callid;
 	}
 
 	/** Get the time the error occurred.
-	 * @return
+	 * @return the error time.
 	 */
 	public long getTime() {
 		return time;